Made progress in TaskBL by adding Get, GetById, and Put functionality.  Made minor change to TaskItem to set UserId to type long instead of int.  Made changes to signature of Get method for IRepository and Repository to account for UserId being changed to long.  Made IUnitOfWork public as it should have been to work with it in .BLL

+++ NotifyWebApi/NotifyWebApi.Repositories/Interfaces/IRepository.cs
@@ -6,7 +6,7 @@ namespace NotifyWebApi.Repositories.Interfaces
 {
     public interface IRepository<TEnt> where TEnt : class
     {
-        TEnt Get(int id);
+        TEnt Get(long id);
         IEnumerable<TEnt> GetAll();
         IEnumerable<TEnt> Find(Expression<Func<TEnt, bool>> predicate);
